backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
redis==5.0.1
PyJWT==2.8.0
Werkzeug==2.3.7
pydantic==1.10.18  
python-multipart==0.0.6
eventlet==0.33.3
email-validator==2.1.0
psycopg2-binary==2.9.9
bcrypt==4.1.2 
cryptography==42.0.0

